pyproject.toml
[project]
name = "2025-11-14-adk-social-data-analytics-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.18.0",
]
